[#1060] Month 표시 모델 5종을 MonthViewModel 에서 떼어 별도 파일로 가른다
- `MonthDisplayModels.swift` 신설 — `WeekDayModel`·`DayCellViewModel`·`WeekRowModel`·`EventMoreModel`·`WeekEventStackViewModel` 과 그 extension 을 옮긴다. 같은 모듈 안이라 접근 제어와 시그니처는 그대로다.
- `MonthViewModel.swift` — 프로토콜과 `MonthViewModelImple` 만 남는다.
